Renamed folders and added test for basic thing items request
